eval/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.12.0
cryptography==2.4.2
cycler==0.10.0
fabric==2.4.0
idna==2.8
invoke==1.2.0
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
pandas==0.24.2
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.3.1
python-dateutil==2.8.0
python3-xlib==0.15
pytz==2018.9
PyYAML==5.1
redis==3.2.1
six==1.12.0
yml==0.0.1
clint==0.5.1